data/requirements.txt
# Akash Network Data Collection Requirements

# HTTP Requests
requests>=2.31.0
aiohttp>=3.9.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Database
pymongo>=4.6.0  # For MongoDB storage
psycopg2-binary>=2.9.9  # For PostgreSQL storage

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Logging
loguru>=0.7.2

# Rate limiting
ratelimit>=2.2.1

# Date/Time handling
python-dateutil>=2.8.2

# JSON handling
orjson>=3.9.10

# Cosmos SDK / gRPC (optional, for advanced features)
grpcio>=1.60.0
protobuf>=4.25.0

# Data validation
pydantic>=2.5.0

# Scheduling (for automated collection)
schedule>=1.2.0
apscheduler>=3.10.4

# CLI tools
click>=8.1.7
rich>=13.7.0  # For beautiful terminal output
